fix bug raised when the fleetmonitoring api cannot return meta data
- update httpstore open_json
